@@ -11,12 +11,12 @@
   "devDependencies": {
     "@types/jquery": "^3.2.12",
     "awesome-typescript-loader": "^3.2.3",
-    "typescript": "^2.5.2",
-    "webpack": "^3.6.0"
+    "typescript": "^2.6.2",
+    "webpack": "^3.10.0"
   },
   "dependencies": {
-    "css-loader": "^0.28.7",
-    "jquery": "1.x",
+    "css-loader": "^0.28.9",
+    "jquery": "latest",
     "jquery-tab": "^1.0.3",
     "style-loader": "^0.18.2"
   }
