@@ -1,4 +1,4 @@
-import $= require('jquery');
+import $ = require('jquery');
 import 'jquery-tab';
 
 $('.tab1').tab();
